evaluators/ragas/pyproject.toml
[tool.poetry]
name = "langevals-ragas"
version = "0.1.0"
description = "LangEvals Ragas evaluator"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
langevals-core = { path = "../../langevals_core", develop = true }
ragas = "0.1.0"
langchain-openai = "^0.0.8"
langchain_community = "^0.0.27"
datasets = "^2.18.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
python-dotenv = "1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
